requirements.txt
freewili
google-genai
elevenlabs
Flask
psutil
python-dotenv
pydantic
